[REFACTOR] Moved the error processor to it's own class.
